# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MobiDuka Retail POS API Blueprint",
    "description": "Multi-tenant, offline-first production backend engine supporting Flutter mobile clients and Kenyan retail ecosystems (Cash tracking, M-Pesa records, local synchronization syncing).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current app origin API router"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT issued by /auth/login. Use the token as Authorization: Bearer <token>."
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Multi-Factor Authentication Entry Gateway",
        "description": "Handles standard credential authorization maps for owners and fast PIN code verification sequences for shift cashiers.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Use the owner password example or the cashier PIN example. Do not send password and pin together.",
                "properties": {
                  "identifier": { "type": "string" },
                  "password": { "type": "string" },
                  "pin": { "type": "string" }
                }
              },
              "examples": {
                "ownerPassword": {
                  "summary": "Seeded owner password login",
                  "value": { "identifier": "owner@mobiduka.com", "password": "OwnerPass123" }
                },
                "cashierPin": {
                  "summary": "Seeded cashier PIN login",
                  "value": { "identifier": "cashier1", "pin": "2540" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Authentication successful, returns active JWT token parameters." },
          "401": { "description": "Invalid user credentials or mismatched secret keys." }
        }
      }
    },
...
    "/cash/session": {
      "post": {
        "tags": ["Business"],
        "summary": "Cashier Shift Drawer Reconciliation System",
        "description": "Triggers daily shift ignition sequences or processes shift totals with dynamic computed cash variances.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action", "businessId", "userId"],
                "properties": {
                  "action": { "type": "string", "example": "OPEN" },
                  "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e", "description": "Seeded deployed business ID. For another deployment, use businessId from the login response." },
                  "userId": { "type": "string", "format": "uuid", "example": "33d0457d-ecb4-4e60-a05a-223bae2541a4", "description": "Seeded deployed owner ID. For another deployment, use the authenticated user id from login." },
                  "sessionId": { "type": "string", "description": "For CLOSE, use the sessionId returned by the OPEN action." },
                  "openingCash": { "type": "number", "example": 5000.0 },
                  "closingCash": { "type": "number", "example": 14200.0 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Session closed out cleanly, returns computed float variances." },
          "201": { "description": "New open shift session generated." }
        }
      }
    },
...
    "/categories": {
      "get": { "tags": ["Inventory"], "summary": "List product categories for a business", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }], "responses": { "200": { "description": "Categories returned with emojis and product counts." }, "401": { "description": "Authenticated business context is required." } } },
      "post": { "tags": ["Inventory"], "summary": "Create a product category", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "name": { "type": "string" }, "emoji": { "type": "string" }, "parentCategoryId": { "type": "string", "format": "uuid" } } } } } }, "responses": { "201": { "description": "Category created." }, "400": { "description": "Invalid category data." }, "401": { "description": "Authenticated business context is required." }, "409": { "description": "Category already exists." } } },
      "patch": { "tags": ["Inventory"], "summary": "Update a product category", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["id", "businessId"], "properties": { "id": { "type": "string", "format": "uuid" }, "businessId": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" }, "name": { "type": "string" }, "emoji": { "type": "string" }, "parentCategoryId": { "type": "string", "format": "uuid" } } } } } }, "responses": { "200": { "description": "Category updated." }, "400": { "description": "Invalid category data." }, "404": { "description": "Category not found." }, "409": { "description": "Category name already exists." } } },
      "delete": { "tags": ["Inventory"], "summary": "Delete an empty product category", "parameters": [{ "name": "businessId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid", "example": "0e722844-15f4-4807-8a69-9614e448f21e" } }, { "name": "id", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }], "responses": { "200": { "description": "Category deleted." }, "400": { "description": "Required identifiers are missing." }, "404": { "description": "Category not found." }, "409": { "description": "Category still has products or child categories." } } }
    },
...
  }
}